test: migrates main ability tests to ts
